refactor: change to spawn waitzone with player
1) 플레이어 웨이팅 존에 생성되도록 로직 수정 및 개선
