--- Assets/01.KyungTaek/02.Prefabs_KT/HealFlower2.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1ecdc3f1b0ab946cc9ac70aeafcf4986
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
