--- Assets/01.KyungTaek/02.Prefabs_KT/JumpFlower2.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 63426ef11d21047cdb6c7c65f79f6928
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
