+++ Assets/Resources/Prefabs/Maps/Object.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 74e5e54e248d81e498a1aae7f3feef72
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
