+++ Assets/Resources/Prefabs/Maps/Object/HealFlower.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1ecdc3f1b0ab946cc9ac70aeafcf4986
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
