--- Assets/Resources/Prefabs/Unit/PlayerPiora.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b5324f7483dc84ee08a920f7a7469afc
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
