+++ Assets/Scripts/Controller/TeamController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bbe2c119aa114e74aaa6302cf3fc3511
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
