+++ Assets/Scripts/Managers/Contents.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a5bad538f90a80345bb7e85c7fd9cf6f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
