+++ ProjectSettings/TagManager.asset
@@ -5,6 +5,8 @@ TagManager:
   serializedVersion: 2
   tags:
   - Bush
+  - HealFlowerSpawnPoint
+  - ExplosionFlowerSpawnPoint
   layers:
   - Default
   - TransparentFX
